pyproject.toml
[project]
name = "drf-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "pytest>=8.4.2",
]
